Fix - 상태값의 이전값 추가
